requirements.txt
streamlit>=1.30.0
requests>=2.31.0
openai>=1.0.0
pypdf>=3.0.0
python-docx>=1.0.0
python-dotenv
